requirements.txt
telethon>=1.28.5
sqlalchemy>=2.0.0
click>=8.1.3
cryptography>=41.0.0
pysqlcipher3>=1.2.0
python-dotenv>=1.0.0
schedule>=1.2.0
